configurator/requirements.txt
flask
python-dotenv
waitress
